# perf: remove some helpless debug_print_data