… polyjuice-tests/src/test_cases/evm-contracts/SimpleWallet.bin