Since debug_print_data takes large runtime overhead, we should reduce
this kinds of `debug_print_data` with long data_length.

![image](https://user-images.githubusercontent.com/1297478/128040495-00211f2b-c7ca-470c-8638-35a6ca71c5e1.png)