@Flouse 2021-08-03
## Benchmarking Deploy_Contract/SimpleStorage: Analyzing

### Before
```
Deploy_Contract/SimpleStorage
                        time:   [21.260 ms 21.752 ms 22.410 ms]
```

### After
[Deploy_Contract/SimpleStorage](https://github.com/nervosnetwork/godwoken-polyjuice/pull/70/checks?check_run_id=3232227694#step:6:2143)
                        time:   [14.157 ms 14.596 ms 15.351 ms]